code and tests to allow spark layer to query using protobufs instead of path to query file

+++ example/java/TestGenomicsDBDataSourceV2.java
@@ -29,6 +29,9 @@ import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 
+import org.genomicsdb.spark.GenomicsDBInput;
+import org.genomicsdb.model.GenomicsDBExportConfiguration;
+
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -38,6 +41,8 @@ import java.util.*;
 
 import static org.apache.spark.sql.functions.bround;
 import static org.apache.spark.sql.functions.col;
+
+import com.googlecode.protobuf.format.JsonFormat;
 
 public final class TestGenomicsDBDataSourceV2 {
 
@@ -110,23 +115,26 @@ public final class TestGenomicsDBDataSourceV2 {
     }
   }
 
-  public static void main(final String[] args) throws IOException {
+  public static void main(final String[] args) throws IOException,
+         org.json.simple.parser.ParseException {
     LongOpt[] longopts = new LongOpt[5];
     longopts[0] = new LongOpt("loader", LongOpt.REQUIRED_ARGUMENT, null, 'l');
     longopts[1] = new LongOpt("query", LongOpt.REQUIRED_ARGUMENT, null, 'q');
     longopts[2] = new LongOpt("vid", LongOpt.REQUIRED_ARGUMENT, null, 'v');
     longopts[3] = new LongOpt("hostfile", LongOpt.REQUIRED_ARGUMENT, null, 'h');
     longopts[4] = new LongOpt("spark_master", LongOpt.REQUIRED_ARGUMENT, null, 's');
+    longopts[5] = new LongOpt("use-query-protobuf", LongOpt.NO_ARGUMENT, null, 'p');
 
-    if (args.length != 10) {
+    if (args.length < 10) {
       System.err.println(
           "Usage:\n\t--loader <loader.json> --query <query.json> --vid <vid.json> "
-              + "--hostfile <hostfile> --spark_master <sparkMaster>");
+              + "--hostfile <hostfile> --spark_master <sparkMaster> --use-query-protobuf");
       System.exit(-1);
     }
     String loaderFile, queryFile, hostfile, vidMapping, sparkMaster, jarDir;
+    boolean useQueryProtobuf = false;
     loaderFile = queryFile = hostfile = sparkMaster = vidMapping = "";
-    Getopt g = new Getopt("TestGenomicsDBSparkHDFS", args, "l:q:h:s:v:", longopts);
+    Getopt g = new Getopt("TestGenomicsDBSparkHDFS", args, "l:q:h:s:v:p", longopts);
     int c = -1;
     String optarg;
 
@@ -147,6 +155,9 @@ public final class TestGenomicsDBDataSourceV2 {
         case 's':
           sparkMaster = g.getOptarg();
           break;
+        case 'p':
+          useQueryProtobuf = true;
+          break;
         default:
           System.err.println("Unknown command line option " + g.getOptarg());
           System.exit(-1);
@@ -157,32 +168,55 @@ public final class TestGenomicsDBDataSourceV2 {
     Path dstdir = Paths.get("").toAbsolutePath();
     Path qSrc = Paths.get(queryFile);
     Path lSrc = Paths.get(loaderFile);
-    File qDstFile = File.createTempFile("query", ".json", new File(dstdir.toString()));
-    qDstFile.deleteOnExit();
+    File qDstFile = null;
+    if (!useQueryProtobuf) {
+      qDstFile = File.createTempFile("query", ".json", new File(dstdir.toString()));
+      qDstFile.deleteOnExit();
+    }
     File lDstFile = File.createTempFile("loader", ".json", new File(dstdir.toString()));
     lDstFile.deleteOnExit();
-    Files.copy(qSrc, qDstFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+    if (!useQueryProtobuf) {
+      Files.copy(qSrc, qDstFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+    }
     Files.copy(lSrc, lDstFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
 
     SparkContext sc = spark.sparkContext();
-    sc.addFile(qDstFile.getName());
+    if (!useQueryProtobuf) {
+      sc.addFile(qDstFile.getName());
+    }
     sc.addFile(lDstFile.getName());
 
     StructType schema = null;
     try {
-      schema = getSchemaFromQuery(qDstFile, vidMapping);
+      schema = getSchemaFromQuery(new File(queryFile), vidMapping);
     } catch (ParseException e) {
       e.printStackTrace();
     }
 
-    Dataset<Row> variants =
-        spark.read()
-            .format("org.genomicsdb.spark.GenomicsDBDataSourceV2")
-            .schema(schema)
-            .option("genomicsdb.input.loaderjsonfile", lDstFile.getName())
-            .option("genomicsdb.input.queryjsonfile", qDstFile.getName())
-            .option("genomicsdb.input.mpi.hostfile", hostfile)
-            .load();
+    Dataset<Row> variants;
+    if (!useQueryProtobuf) {
+      variants =
+          spark.read()
+              .format("org.genomicsdb.spark.GenomicsDBDataSourceV2")
+              .schema(schema)
+              .option("genomicsdb.input.loaderjsonfile", lDstFile.getName())
+              .option("genomicsdb.input.queryjsonfile", qDstFile.getName())
+              .option("genomicsdb.input.mpi.hostfile", hostfile)
+              .load();
+    }
+    else {
+      GenomicsDBExportConfiguration.ExportConfiguration.Builder builder =
+          GenomicsDBInput.getExportConfigurationFromJsonFile(queryFile);
+      String pbString = JsonFormat.printToString(builder.build());
+      variants =
+          spark.read()
+              .format("org.genomicsdb.spark.GenomicsDBDataSourceV2")
+              .schema(schema)
+              .option("genomicsdb.input.loaderjsonfile", lDstFile.getName())
+              .option("genomicsdb.input.queryprotobuf", pbString)
+              .option("genomicsdb.input.mpi.hostfile", hostfile)
+              .load();
+    }
 
     String tempDir = "./" + UUID.randomUUID().toString();
